Remove examples in hipblaslt
Remove deprecated examples
